requirements.txt
# First install setuptools to avoid build issues
setuptools>=65.0.0

# Core packages
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
joblib>=1.1.0
xgboost>=1.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Experiment tracking
mlflow>=2.0.0

# Configuration
pyyaml>=6.0

# Install these separately if needed
# jupyter
# ipykernel
# statsmodels